Extract out history construction and processing stream
